// frontend/src/api/generated/model/productUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * EstacionCafe API
 * API REST de EstacionCafé. Todas las rutas requieren JWT (header `Authorization: Bearer <token>` o cookie `auth_token`) excepto login y logout. Los roles permitidos de cada operación están en su descripción y en `x-roles`.
 * OpenAPI spec version: 2.0.0
 */

/**
 * El precio resultante debe ser mayor al costo resultante
 */
export interface ProductUpdate {
  /**
     * @minLength 1
     * @maxLength 50
     */
  name?: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  description?: string;
  price?: number | string;
  cost?: number | string;
  productTypeId?: number | string;
  active?: boolean;
}